feat: hotels - getting correct ticket status
