+++ src/pages/Dashboard/Hotel/index.js
@@ -2,17 +2,20 @@ import styled from 'styled-components';
 import { Typography } from '@material-ui/core';
 import Warning from '../../../components/WarningMessage';
 import Hotels from '../../../components/Hotels/index';
+import * as useTickets from '../../../hooks/api/useTickets';
 
 export default function Hotel() {
-  const tickets = 'PAID';
+  const { tickets } = useTickets.useTickets();
 
   return (
     <>
       <StyledTypography variant="h4">Escolha de hotel e quarto</StyledTypography>
-      {tickets !== 'PAID' ? (
+      {tickets?.status !== 'PAID' ? (
         <Warning>Você precisa ter confirmado pagamento antes de fazer a escolha da hospedagem</Warning>
       ) : (
-        <Hotels/>
+        <>
+          <Hotels/>
+        </>
       )}
     </>
   );
