+++ src/services/ticketsApi.js
@@ -8,3 +8,13 @@ export async function getTicketTypes(token) {
   });
   return response.data;
 }
+
+export async function getTickets(token) {
+  const response = await api.get('/tickets', {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  });
+
+  return response.data;
+}
